+++ packages/testing/src/chip/index.ts
@@ -99,6 +99,7 @@ export type {
     SubscribeEventOptions,
     SubscribeOptions,
     TimedInteractionOptions,
+    WebRtcIceCandidate,
     WebRtcRequestorApi,
     WebRtcSessionRecord,
     WebRtcSessionSpec,